goodixmoc: Fix print template parsing
The duplicate check was testing against the struct size, but it is
encoded in the packet with a variable lenght. While at it, also remove
the extra size check from identify, as the fingerid parser will do the
proper checking and correctly takes the possible variable length into
account.

Related: #444